made userid an int, which ended up touching a lot

+++ src/mitll/langtest/server/database/instrumentation/IEventDAO.java
@@ -20,7 +20,7 @@ public interface IEventDAO {
 
   SlickSlimEvent getFirstSlim(String language);
 
-  void addPlayedMarkings(long userID, CommonExercise firstExercise);
+  void addPlayedMarkings(int userID, CommonExercise firstExercise);
 
   Number getNumRows(String language);
 }
